requirements.txt
faiss_cpu==1.11.0
fastapi==0.115.14
mysql_connector_repackaged==0.3.1
ollama==0.5.1
openpyxl==3.1.5
protobuf==6.31.1
pydantic==2.11.7
python-dotenv==1.1.1
rich==14.0.0
sentence_transformers==4.1.0
torch==2.7.1
transformers==4.52.4
uvicorn==0.34.3
